+++ tests/xharness/Jenkins/TestTasks/RunTestTask.cs
@@ -10,7 +10,7 @@ using Microsoft.DotNet.XHarness.iOS.Shared.Logging;
 using Microsoft.DotNet.XHarness.iOS.Shared.Utilities;
 
 namespace Xharness.Jenkins.TestTasks {
-	internal abstract class RunTestTask : TestTask
+	internal abstract class RunTestTask : JenkinsTestTask
 	{
 		protected IProcessManager ProcessManager { get; }
 		IResultParser ResultParser { get; } = new XmlResultParser ();
