+++ tests/xharness/TestTasks/IAcquiredResource.cs
@@ -1,0 +1,7 @@
+﻿using System;
+namespace Xharness.TestTasks {
+	public interface IAcquiredResource : IDisposable
+	{
+		Resource Resource { get; }
+	}
+}
